Adds the postgresql-data-k8s charm
This charm can be related to a postgresql-k8s charm (using the db-admin relation), and
this charm will execute the SQL tar dump given as a resource.

If the refresh-period (minutes) is non-zero, this charm will reexecute the SQL tar dump after
that number of minutes have passed since the last update.

This charm can be configured with the db-user config option, which will be the new owner of the
newly created data (otherwise, consumers of that database will encounter Permission Denied
errors).