Updates the charm to accept an SQL dump URL instead
Having an sql-dump-url config option will make this charm easier
to deploy and include in bundles. Plus, it will make it easier to
update the URL.

The SQL dump file is downloaded each time to avoid the scenario in
which it was changed at the destination. If the downloaded file is
a .gz file, this charm will also unpack it before using it.