Adds upgrade event hook
If a Pod is rebuilt in Kubernetes for any reason, it won't trigger the
on install hook anymore, meaning that the new Pod will not have the
postgresql-client installed, which can cause the charm to misbehave.

An upgrade hook is triggered when the Pod is rebuilt, so we're adding
a hook for that, in which we'll install the postgresql-client. Note that
the upgrade event is no triggered when first deploying the charm, which
is why we're keeping that event hook.